MCP: [FEAT] Remove MergedQueue from scheduler controller
Controller is the base modul of scheduler. The mergedqueue is no longer
required and I removed it. Now the queues themselves are responsible for
fitting hosts. Additionally we need a priority queue to make sure tests
belonging to the same scenario are scheduled close together. This patch brings
this priority queue.