MCP: [FEAT] no fitting on already tried queues
Previously when the Scheduler tried to fit a queue and was not able to do so
it asked the algorithm what the next queue to try fitting would be. Because of
bandwith this could well be the one that said "no fitting" right before and in
fact was many many times before all queues where tried. Now a queue that says
"I don't have a testrequest that fits any of your free hosts" is not asked
again in this round. Thus, no more that $number_of_queues fitting tries are
executed.