MCP: [FEAT] no penalty if fitting queue is not first choice
If the queue chosen first by the algorithm does not have a fitting we
ask other queues until a fitting is found or all queues got their
chance. This patch prevents increasing the runcount of all but the first
queue if a fitting is found. Thus, the other queues won't receive a
penalty for increasing the utilisation of hosts. Fixes BZ-690.