MCP: [FIX] check more carefully whether SSH reboot succeeded
Actually, we do not have a good way to be sure that ssh reboot
succeeded. Therefore we try to parse the output generated by
remote. Before we concluded a successful reboot from finding 'ogin'
somewhere in the exchange for logging in. Now we analyse the output of
reboot too and search it for "System is going down for reboot". This
should match every linux system.