[FIX] improve race conditions in timeout tests
During timeout computations there are continuous
probes of current time which to track from the
outside (in a testscript) would require some
intrusive changes in the module itself.

We now cover the most typical race conditions where
a timestamp is taken just before a second changes
and shortly after the change of the second and
therefore produce off-by-one mistakes. We have
some -1/+1 adjustment for each such a probe
that we know happens inside.

That reduces the problems to the coarse cases
of a machine under heavy load where few lines of code
take longer than a second.