temare for producer test is now a shell script
For unit testing the Temare producer a test script is called. This test
temare script was perl and tried to use YAML. It called /usr/bin/env
perl which may not be the perl used for the test and thus YAML may be
unavailable. Now the test temare is a shell script that only does an
echo. This should be available on all platforms.