[FIX] Fixes BZ-#1036, timeout issues for testprograms
When a testprogram hangs up the test machine no end-testprogram and no
end-testing will be sent. In such a case MCP did try to set the timeout
to the next testprogram in this PRC without checking whether such a
testprogram exists. For an inexisting testprogram a default timeout was
used. This patch fixes this behaviour.