[FEAT] producer in virt preconditions
This patch implements producer evaluation in virt
preconditions. Everything that is a precondition inside the virt
precondition can be a producer now. This means it is not a template
mechanism. For example you need to know in advance how many guest you
need and can not let a producer decide this. Elements that allow only
one precondition (e.g. guest root, config, host root, ...) will always
use the first precondition returned by the producer. If the producer
delivers additional preconditions they will be ignored. When a list is
allowed (e.g. host preconditions, testrun_list, ...) all preconditions
delivered by the producer will be added. The producer precondition is
substituted by the new produced preconditions, so order of the list is
preserved.