[FIX] substitute ssh_cmd with ssh
Net::SSH::ssh_cmd dies if anything is printed on STDERR. Since we have
valid text to be printed there (e.g. "Warning: Permanently added
'hostname' (RSA) to the list of known hosts.") we need to use a function
that does allow this text. Therefore, it's now Net::SSH::ssh.