[FIX] handling of 'quit' message (freehost)
The message 'quit' primarily comes from
the CLI via

  tapper-testrun freehost --name HOSTNAME

and is meant to declare the current testrun
as finished so the host can be reused for the
next queued testruns.

Originally it was handled by just setting
the current_state of the testrun to 'finished'.
Now we also set the state of all sub-PRCs to
'finished' because the state handling seems
to assume that.