[RFK] default installer grub now configurable
We now allow to set the default grub entry
written for Installer to be configured in
/etc/tapper.cfg.

Like this

mcp:
  installer:
    default_grub: |
      default 0
      timeout 2
      title Test
           tftpserver
           kernel  root=/dev/nfs ro ... $TAPPER_OPTIONS

There is still a default value hardcoded
internally for backcompat reasons. So only if
you want to have something else than it was in
the past then add such an entry.